prompt_to_image_generation/requirements.txt
flask
flask-mysqldb
mysql-connector-python
diffusers
torch
transformers
pillow
python-dotenv
werkzeug